CRITICAL Change: User Model and REST Calls
*users dictionary type changed (AppContext.swift)
*User latitude and longitude attributes changed from float to double.
Make sure that this doesn’t break the radar as it seems to accept float
values. Revert back to float if it does break. (User.swift)

+++ TeamTracking/TeamTracking/User.swift
@@ -10,8 +10,8 @@ import UIKit
 
 class User: NSObject {
     var name = String()
-    var latitude = Float()
-    var longitude = Float()
+    var latitude = Double()
+    var longitude = Double()
     var distance = Int()
     var available = Bool()
     var color = UIColor()
